+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <!-- SPDX-License-Identifier: Apache-2.0 -->
 # Changelog
 
+## [7.5.4.0] (Unreleased)
+
+- [usd#2652](https://github.com/Autodesk/arnold-usd/issues/2652) - Add setting for accelerated viewport
 
 ## [7.5.4.0] (Unreleased)